move weft to internal
