+++ cmd/fits-api/charts.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"bytes"
-	"github.com/GeoNet/weft"
+	"github.com/GeoNet/fits/internal/weft"
 	"html/template"
 	"net/http"
 	"os"
