+++ cmd/fits-api/plot_site.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"github.com/GeoNet/fits/internal/ts"
-	"github.com/GeoNet/weft"
+	"github.com/GeoNet/fits/internal/weft"
 	"net/http"
 	"time"
 )
