+++ cmd/fits-api/plot_sites.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/GeoNet/fits/internal/ts"
-	"github.com/GeoNet/weft"
+	"github.com/GeoNet/fits/internal/weft"
 	"net/http"
 	"time"
 )
