+++ internal/weft/wefttest/doc.go
@@ -28,5 +28,5 @@ A minimal test for a web application would look like:
 		}
 	}
 
- */
+*/
 package wefttest
